Encoding and Decoding general structure

+++ api/src/main/java/wtf/file/api/data/Pixel.java
@@ -28,7 +28,7 @@ public interface Pixel {
      *
      * @return a map of {@link ColorChannel} to {@code long} values.
      */
-    Map<ColorChannel, Long> values();
+    Map<ColorChannel, Short> values();
 
     /**
      * Retrieves the numeric value associated with a specific color channel.
@@ -37,7 +37,7 @@ public interface Pixel {
      * @return the value of the specified channel.
      * @throws IllegalArgumentException if the channel is not part of this pixel's {@link ColorSpace}.
      */
-    long valueOf(ColorChannel channel);
+    short valueOf(ColorChannel channel);
 
     /**
      * Converts this pixel to the given {@link ColorSpace}, potentially performing
